// src/models/enums.ts
export enum Gender {
	MALE = 'MALE',
	FEMALE = 'FEMALE',
}

export enum Language {
	ENGLISH = 'ENGLISH',
	SIMPLIFIED_CHINESE = 'SIMPLIFIED CHINESE',
	TRADITIONAL_CHINESE = 'TRADITIONAL CHINESE',
}

export enum PaymentMethod {
	CASH = 'CASH',
	MACHINE = 'MACHINE',
}

export enum Permissions {
	PERMISSION_GET_CUSTOMER = 'GET CUSTOMER PERMISSION',
	PERMISSION_UPDATE_CUSTOMER = 'UPDATE CUSTOMER PERMISSION',
	PERMISSION_ADD_CUSTOMER = 'ADD CUSTOMER PERMISSION',
	PERMISSION_DELETE_CUSTOMER = 'DELETE CUSTOMER PERMISSION',
	PERMISSION_GET_EMPLOYEE = 'GET EMPLOYEE PERMISSION',
	PERMISSION_UPDATE_EMPLOYEE = 'UPDATE EMPLOYEE PERMISSION',
	PERMISSION_ADD_EMPLOYEE = 'ADD EMPLOYEE PERMISSION',
	PERMISSION_DELETE_EMPLOYEE = 'DELETE EMPLOYEE PERMISSION',
	PERMISSION_GET_GIFT_CARD = 'GET GIFT CARD PERMISSION',
	PERMISSION_UPDATE_GIFT_CARD = 'UPDATE GIFT CARD PERMISSION',
	PERMISSION_ADD_GIFT_CARD = 'ADD GIFT CARD PERMISSION',
	PERMISSION_DELETE_GIFT_CARD = 'DELETE GIFT CARD PERMISSION',
	PERMISSION_GET_SCHEDULE = 'GET SCHEDULE PERMISSION',
	PERMISSION_UPDATE_SCHEDULE = 'UPDATE SCHEDULE PERMISSION',
	PERMISSION_ADD_SCHEDULE = 'ADD SCHEDULE PERMISSION',
	PERMISSION_DELETE_SCHEDULE = 'DELETE SCHEDULE PERMISSION',
	PERMISSION_GET_SERVICE = 'GET SERVICE PERMISSION',
	PERMISSION_UPDATE_SERVICE = 'UPDATE SERVICE PERMISSION',
	PERMISSION_ADD_SERVICE = 'ADD SERVICE PERMISSION',
	PERMISSION_DELETE_SERVICE = 'DELETE SERVICE PERMISSION',
	PERMISSION_GET_RESERVATION = 'GET RESERVATION PERMISSION',
	PERMISSION_UPDATE_RESERVATION = 'UPDATE RESERVATION PERMISSION',
	PERMISSION_ADD_RESERVATION = 'ADD RESERVATION PERMISSION',
	PERMISSION_DELETE_RESERVATION = 'DELETE RESERVATION PERMISSION',
	PERMISSION_GET_VIP_PACKAGE = 'GET VIP PACKAGE PERMISSION',
	PERMISSION_UPDATE_VIP_PACKAGE = 'UPDATE VIP PACKAGE PERMISSION',
	PERMISSION_ADD_VIP_PACKAGE = 'ADD VIP PACKAGE PERMISSION',
	PERMISSION_DELETE_VIP_PACKAGE = 'DELETE VIP PACKAGE PERMISSION',
}

export enum Role {
	DEVELOPER = 'SOFTWARE DEVELOPER',
	MANAGER = 'MANAGER',
	RECEPTIONIST = 'RECEPTIONIST',
	STORE_EMPLOYEE = 'STORE EMPLOYEE',
	OTHER = 'OTHER',
}

export enum ServiceColor {
	RED = 'RED',
	BLUE = 'BLUE',
	YELLOW = 'YELLOW',
	GREEN = 'GREEN',
	ORANGE = 'ORANGE',
	PURPLE = 'PURPLE',
	GRAY = 'GRAY',
	BLACK = 'BLACK',
}

export enum TipMethod {
	CASH = 'CASH',
	MACHINE = 'MACHINE',
	HALF = 'HALF',
}